+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/internal/SyncedRealmContext.kt
@@ -15,9 +15,11 @@
  */
 package io.realm.mongodb.internal
 
-import io.realm.Realm
+import io.realm.BaseRealm
+import io.realm.internal.BaseRealmImpl
 import io.realm.internal.RealmImpl
 import io.realm.internal.interop.RealmInterop
+import io.realm.mongodb.sync.SubscriptionSet
 import io.realm.mongodb.sync.SyncConfiguration
 import io.realm.mongodb.sync.SyncSession
 
@@ -30,19 +32,33 @@ import io.realm.mongodb.sync.SyncSession
  * In order to work around the bootstrap problem, all public API entry points that access this
  * class must do so through the [executeInSyncContext] closure.
  */
-internal class SyncedRealmContext(realm: RealmImpl) {
-    internal val session: SyncSession = SyncSessionImpl(realm, RealmInterop.realm_sync_session_get(realm.realmReference.dbPointer))
+internal class SyncedRealmContext<T : BaseRealm>(realm: T) {
+    // TODO For now this can only be a RealmImpl, which is required by the SyncSessionImpl
+    //  When we introduce a public DynamicRealm, this can also be a `DynamicRealmImpl`
+    //  And we probably need to modify the SyncSessionImpl to take either of these two.
+    private val baseRealm = realm as RealmImpl
+    private val dbPointer = baseRealm.realmReference.dbPointer
+    internal val config: SyncConfiguration = baseRealm.configuration as SyncConfiguration
+    internal val session: SyncSession by lazy {
+        SyncSessionImpl(baseRealm, RealmInterop.realm_sync_session_get(dbPointer))
+    }
+    internal val subscriptions: SubscriptionSet<T> by lazy {
+        SubscriptionSetImpl<T>(realm, RealmInterop.realm_sync_get_latest_subscriptionset(dbPointer))
+    }
 }
 
 /**
  * Helper methods that can be used by public API entry points to grant safe access to the
  * [SyncedRealmContext], or otherwise throw an appropriate exception.
  */
-internal fun <T> executeInSyncContext(realm: Realm, block: (context: SyncedRealmContext) -> T): T {
+internal fun <T, R : BaseRealm> executeInSyncContext(realm: R, block: (context: SyncedRealmContext<R>) -> T): T {
+    if (realm.isClosed()) {
+        throw IllegalStateException("This method is not available when the Realm has been closed.")
+    }
     val config = realm.configuration
     if (config is SyncConfiguration) {
-        if (realm is RealmImpl) {
-            val context: SyncedRealmContext = initSyncContextIfNeeded(realm)
+        if (realm is BaseRealmImpl) {
+            val context: SyncedRealmContext<R> = initSyncContextIfNeeded(realm)
             return block(context)
         } else {
             // Should never happen. Indicates a problem with our internal architecture.
@@ -54,19 +70,19 @@ internal fun <T> executeInSyncContext(realm: Realm, block: (context: SyncedRealm
     }
 }
 
-private fun initSyncContextIfNeeded(realm: RealmImpl): SyncedRealmContext {
+private fun <T : BaseRealm> initSyncContextIfNeeded(realm: T): SyncedRealmContext<T> {
     // INVARIANT: `syncContext` is only ever set once, and never to `null`.
     // This code works around the fact that `Mutex`'s can only be locked inside suspend functions on
     // Kotlin Native.
-    val syncContext = realm.syncContext
+    val syncContext = (realm as RealmImpl).syncContext
     return if (syncContext.value != null) {
-        syncContext.value!! as SyncedRealmContext
+        syncContext.value!! as SyncedRealmContext<T>
     } else {
         // Worst case, two SyncedRealmContext will be created and one of them will thrown
         // away. As long as SyncedRealmContext is cheap to create, this should be fine. If, at
         // some point, it start having too much state, we can consider making `lazy` properties
         // inside the class to defer the construction cost.
-        syncContext.compareAndSet(null, SyncedRealmContext(realm))
-        syncContext.value!! as SyncedRealmContext
+        syncContext.compareAndSet(null, SyncedRealmContext<T>(realm))
+        syncContext.value!! as SyncedRealmContext<T>
     }
 }