+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/sync/SyncConfiguration.kt
@@ -40,6 +40,36 @@ import io.realm.mongodb.internal.UserImpl
 import kotlin.reflect.KClass
 
 /**
+ * This enum determines how Realm sync data with the server.
+ *
+ * The server must be configured for the selected way, otherwise an error will be
+ * reported to [SyncConfiguration.errorHandler] when the Realm connects to the server
+ * for the first time.
+ */
+public enum class SyncMode {
+    /**
+     * Partition-based Sync. Data is selected for synchronization based on a _partition key_,
+     * which is a property that must be set on all objects. Server objects that
+     * match a given _partition value_ are then synchronized to the device.
+     *
+     * @see https://www.mongodb.com/docs/atlas/app-services/sync/data-access-patterns/partitions/
+     */
+    PARTITION_BASED,
+
+    /**
+     * Flexible Sync. Data is selected for synchronization based on one or more queries which are
+     * stored in a [SubscriptionSet]. All server objects that match one or more queries are then
+     * synchronized to the device.
+     *
+     * @see https://www.mongodb.com/docs/atlas/app-services/sync/data-access-patterns/flexible-sync/
+     */
+    FLEXIBLE
+}
+
+// TODO https://github.com/realm/realm-kotlin/issues/840
+internal typealias InitialSubscriptionsCallback = MutableSubscriptionSet.(realm: Realm) -> Unit
+
+/**
  * A [SyncConfiguration] is used to setup a Realm Database that can be synchronized between
  * devices using MongoDB Realm.
  *
@@ -64,12 +94,27 @@ public interface SyncConfiguration : Configuration {
     public val errorHandler: SyncSession.ErrorHandler?
 
     /**
+     * The mode of synchronization for this Realm
+     */
+    public val syncMode: SyncMode
+
+    // /**
+    //  * TODO https://github.com/realm/realm-kotlin/issues/840
+    //  */
+    // public val initialSubscriptionsCallback: InitialSubscriptionsCallback?
+    //
+    // /**
+    //  * TODO https://github.com/realm/realm-kotlin/issues/840
+    //  */
+    // public val rerunInitialSubscriptions: Boolean
+
+    /**
      * Used to create a [SyncConfiguration]. For common use cases, a [SyncConfiguration] can be
      * created using the [SyncConfiguration.with] function.
      */
     public class Builder private constructor(
         private var user: User,
-        private var partitionValue: PartitionValue,
+        private var partitionValue: PartitionValue?,
         schema: Set<KClass<out RealmObject>>,
     ) : Configuration.SharedBuilder<SyncConfiguration, Builder>(schema) {
 
@@ -78,19 +123,72 @@ public interface SyncConfiguration : Configuration {
         protected override var name: String? = null
 
         private var errorHandler: SyncSession.ErrorHandler? = null
+        private var initialSubscriptionHandler: InitialSubscriptionsCallback? = null
+        private var rerunInitialSubscriptions: Boolean = false
 
+        /**
+         * Creates a [SyncConfiguration.Builder] for Flexible Sync. Flexible Sync must be enabled
+         * on the server for this to work.
+         *
+         * @param user user used to access server side data. This will define which data is
+         * available from the server.
+         * @param schema the classes of the schema. The elements of the set must be direct class
+         * literals.
+         * @see https://www.mongodb.com/docs/realm/sync/data-access-patterns/flexible-sync/
+         */
+        public constructor(
+            user: User,
+            schema: Set<KClass<out RealmObject>>
+        ) : this(user, null as PartitionValue?, schema)
+
+        /**
+         * Creates a [SyncConfiguration.Builder] for Partition-Based Sync. Partition-Based Sync
+         * must be enabled on the server for this to work.
+         *
+         * @param user user used to access server side data. This will define which data is
+         * available from the server.
+         * @param partitionValue the partition value to use data from. The server must have been
+         * configured with a Int partition key for this to work.
+         * @param schema the classes of the schema. The elements of the set must be direct class
+         * literals.
+         * @see https://www.mongodb.com/docs/realm/sync/data-access-patterns/partitions/
+         */
         public constructor(
             user: User,
             partitionValue: Int?,
             schema: Set<KClass<out RealmObject>>
         ) : this(user, PartitionValue(partitionValue), schema)
 
+        /**
+         * Creates a [SyncConfiguration.Builder] for Partition-Based Sync. Partition-Based Sync
+         * must be enabled on the server for this to work.
+         *
+         * @param user user used to access server side data. This will define which data is
+         * available from the server.
+         * @param partitionValue the partition value to use data from. The server must have been
+         * configured with a Long partition key for this to work.
+         * @param schema the classes of the schema. The elements of the set must be direct class
+         * literals.
+         * @see https://www.mongodb.com/docs/realm/sync/data-access-patterns/partitions/
+         */
         public constructor(
             user: User,
             partitionValue: Long?,
             schema: Set<KClass<out RealmObject>>
         ) : this(user, PartitionValue(partitionValue), schema)
 
+        /**
+         * Creates a [SyncConfiguration.Builder] for Partition-Based Sync. Partition-Based Sync
+         * must be enabled on the server for this to work.
+         *
+         * @param user user used to access server side data. This will define which data is
+         * available from the server.
+         * @param partitionValue the partition value to use data from. The server must have been
+         * configured with a String partition key for this to work.
+         * @param schema the classes of the schema. The elements of the set must be direct class
+         * literals.
+         * @see https://www.mongodb.com/docs/realm/sync/data-access-patterns/partitions/
+         */
         public constructor(
             user: User,
             partitionValue: String?,
@@ -139,6 +237,14 @@ public interface SyncConfiguration : Configuration {
             this.name = name
         }
 
+        internal fun initialSubscriptions(
+            rerunOnOpen: Boolean = false,
+            initialSubscriptionBlock: InitialSubscriptionsCallback
+        ): Builder = apply {
+            this.rerunInitialSubscriptions = rerunOnOpen
+            this.initialSubscriptionHandler = initialSubscriptionBlock
+        }
+
         override fun build(): SyncConfiguration {
             val allLoggers = userLoggers.toMutableList()
             // TODO This will not remove the system logger if it was added in AppConfiguration and
@@ -171,6 +277,11 @@ public interface SyncConfiguration : Configuration {
                 }
             }
 
+            // ObjectStore uses a different default value for Flexible Sync than we want,
+            // so inject our default name if no user provided name was found
+            if (partitionValue == null && name == null) {
+                name = Realm.DEFAULT_FILE_NAME
+            }
             val fullPathToFile = getAbsolutePath(name)
             val fileName = fullPathToFile.substringAfterLast(PATH_SEPARATOR)
             val directory = fullPathToFile.removeSuffix("$PATH_SEPARATOR$fileName")
@@ -195,7 +306,9 @@ public interface SyncConfiguration : Configuration {
                 baseConfiguration,
                 partitionValue,
                 user as UserImpl,
-                errorHandler!! // It will never be null: either default or user-provided
+                errorHandler!!, // It will never be null: either default or user-provided
+                initialSubscriptionHandler,
+                rerunInitialSubscriptions
             )
         }
 
@@ -204,10 +317,10 @@ public interface SyncConfiguration : Configuration {
             // configuration which at this point we don't yet have, so we have to create a
             // temporary one so that we can return the actual path to a sync Realm using the
             // realm_app_sync_client_get_default_file_path_for_realm function below
-            val absolutePath: String = RealmInterop.realm_sync_config_new(
-                (user as UserImpl).nativePointer,
-                partitionValue.asSyncPartition()
-            ).let { auxSyncConfig ->
+            val absolutePath: String = when (partitionValue == null) {
+                true -> RealmInterop.realm_flx_sync_config_new((user as UserImpl).nativePointer)
+                false -> RealmInterop.realm_sync_config_new((user as UserImpl).nativePointer, partitionValue!!.asSyncPartition())
+            }.let { auxSyncConfig ->
                 RealmInterop.realm_app_sync_client_get_default_file_path_for_realm(
                     (user as UserImpl).app.nativePointer,
                     auxSyncConfig,
@@ -225,6 +338,21 @@ public interface SyncConfiguration : Configuration {
     }
 
     public companion object {
+
+        /**
+         * Creates a sync configuration for Flexible Sync with default values for all
+         * optional configuration parameters.
+         *
+         * Flexible Sync uses a concept called subscription sets to define which data gets
+         * uploaded and downloaded to the device. See [SubscriptionSet] for more information.
+         *
+         * @param user the [User] who controls the realm.
+         * @param schema the classes of the schema. The elements of the set must be direct class literals.
+         * @throws IllegalArgumentException if the user is not valid and logged in.
+         * @see https://www.mongodb.com/docs/atlas/app-services/sync/data-access-patterns/flexible-sync/
+         */
+        public fun with(user: User, schema: Set<KClass<out RealmObject>>): SyncConfiguration =
+            Builder(user, schema).build()
 
         /**
          * Creates a sync configuration for Partition-based Sync with default values for all